+++ spec/dummy_app/app/models/thing.rb
@@ -1,7 +1,10 @@
 # frozen_string_literal: true
 
 class Thing < ActiveRecord::Base
-  has_paper_trail
+  has_paper_trail versions: {
+    extend: PrefixVersionsInspectWithCount,
+    scope: -> { order("id desc") }
+  }
 
   if ActiveRecord.gem_version >= Gem::Version.new("5.0")
     belongs_to :person, optional: true